+++ modules/app.rb
@@ -1,148 +1,33 @@
-require_relative 'book'
-require_relative 'classroom'
-require_relative 'decorator'
-require_relative 'nameable'
-require_relative 'person'
-require_relative 'rental'
-require_relative 'student'
-require_relative 'teacher'
+require_relative 'library_management'
 
 EXIT_MESSAGE = "\n\nThank you for using this app\n\n".freeze
 
 class App
+  MENU_OPTIONS = {
+    '1' => :list_books,
+    '2' => :list_people,
+    '3' => :create_person,
+    '4' => :create_book,
+    '5' => :create_rental,
+    '6' => :list_rentals_for_person
+  }.freeze
+
   def initialize
-    @books = []
-    @people = []
-    @rentals = []
-  end
-
-  def list_books
-    if @books.empty?
-      puts "\nNo books to display. Create a book!"
-    else
-      puts "\nBooks:"
-      @books.each_with_index { |book, i| puts "#{i}) #{book.title} by #{book.author}" }
-    end
-  end
-
-  def list_people
-    if @people.empty?
-      puts "\nNo people to display. Create a person!"
-    else
-      puts "\nPeople:"
-      @people.each_with_index do |person, i|
-        person_type = person.is_a?(Teacher) ? 'Teacher' : 'Student'
-        specialization = person.is_a?(Teacher) ? "specialization: #{person.specialization}" : nil
-        puts "#{i}) [#{person_type}] name: #{person.name}, ID: #{person.id}, age: #{person.age}, #{specialization}"
-      end
-    end
-  end
-
-  def create_person
-    print "\nDo you want to create a student (1) or a teacher (2)? [Input the number]: "
-    person_type = gets.chomp.to_i
-
-    case person_type
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      puts 'Invalid option. Please choose (1) for Student or (2) for Teacher.'
-    end
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    teacher = Teacher.new(age, specialization, name)
-    @people << teacher
-    puts "Teacher #{teacher.name} created successfully!"
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    student = Student.new(age, name)
-    @people << student
-    puts "Student #{student.name} created successfully!"
-  end
-
-  def create_book
-    print "\nTitle: "
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @books << book
-    puts 'Book created successfully!'
-  end
-
-  def create_rental
-    list_people
-    print 'Select a person from the following list by number (not id): '
-    person_id = gets.chomp.to_i
-    person = @people[person_id]
-
-    unless person
-      puts 'Person not found!'
-      return
-    end
-
-    list_books
-    print 'Select a book from the following list by number: '
-    book_id = gets.chomp.to_i
-    book = @books[book_id]
-
-    unless book
-      puts 'Book not found!'
-      return
-    end
-
-    rental = person.add_rental(book)
-    @rentals << rental
-
-    puts "#{person.name} (ID: #{person.id}) has successfully rented #{book.title}!"
-  end
-
-  def list_rentals_for_person
-    list_people
-    print "\nEnter the ID of the person to list rentals: "
-    person_id = gets.chomp.to_i
-    person = @people.find { |p| p.id == person_id }
-
-    unless person
-      puts 'Person not found!'
-      return
-    end
-
-    person_type = person.is_a?(Teacher) ? 'Teacher' : 'Student'
-
-    puts "Rentals for [#{person_type}] #{person.name} (ID: #{person.id}):"
-    person.rentals.each do |rental|
-      puts "Date: #{rental.date}, Book: \"#{rental.book.title}\" by #{rental.book.author}"
-    end
+    @library_system = LibraryManagement.new
   end
 
   def start
-    loop do
-      display_menu
-      choice = user_choice
-      if choice.zero?
-        puts EXIT_MESSAGE
-        break
-      end
-      process_choice(choice)
-    rescue Interrupt
-      puts EXIT_MESSAGE
-      break
+    display_menu
+    choice = user_choice
+
+    if MENU_OPTIONS.key?(choice)
+      send(MENU_OPTIONS[choice])
+    else
+      puts 'Invalid option. Please try again.'
     end
+  rescue Interrupt
+    puts EXIT_MESSAGE
+    break
   end
 
   private
@@ -158,27 +43,32 @@ class App
     puts '0 - Exit'
   end
 
-  def process_choice(choice)
-    case choice
-    when 1
-      list_books
-    when 2
-      list_people
-    when 3
-      create_person
-    when 4
-      create_book
-    when 5
-      create_rental
-    when 6
-      list_rentals_for_person
-    else
-      puts 'Invalid option. Please try again.'
-    end
-  end
-
   def user_choice
     print '>> '
-    gets.chomp.to_i
+    gets.chomp.strip
+  end
+
+  def list_books
+    @library_system.list_books
+  end
+
+  def list_people
+    @library_system.list_people
+  end
+
+  def create_person
+    @library_system.create_person
+  end
+
+  def create_book
+    @library_system.create_book
+  end
+
+  def create_rental
+    @library_system.create_rental
+  end
+
+  def list_rentals_for_person
+    @library_system.list_rentals_for_person
   end
 end
